requirements.txt
mmocr==0.6.0
numpy==1.20.1
opencv_python==4.5.5.62
Pillow==9.1.0
streamlit==1.4.0
